// src/app/model/miscellaneous.model.ts
export interface Miscellaneous {
  id: string;
  user_id: string;
  group_id: string;
  amount: string;
  start_date: string;
  end_date: string;
  loan_emi: string;
  title: string;
  description: string;
  status: string;
  payment_method: string;
  paid_status: string;
  created_at: string;
  tenure: string;
  sno: number;
}